requirements_dev.txt
# Development environment pip requirements file
# Install via 'pip install -r requirements_dev.txt'

Pillow
jinja2
simplejson
Flask
SQLAlchemy
Flask-SQLAlchemy
unidecode
pycrypto

